// src/models/account.ts
export interface AccountResponse {
    status: number
    data: {
        puuid: string
        region: string
        account_level: number
        name: string
        tag: string
        card: {
            small: string
            large: string
            wide: string
            id: string
        }
        last_update: string
        last_update_raw: number
    }
}